python-sidecar/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
requests==2.32.3
numpy==2.2.3
pandas==2.2.3
scikit-learn==1.6.1
joblib==1.4.2
